menu

+++ Assets/Scripts/PlayerSelector.cs
@@ -4,8 +4,12 @@ using UnityEngine;
 
 public class PlayerSelector : MonoBehaviour
 {
+    private int gold;
     public GameObject nextButton;
     public GameObject backwardButton;
+    public GameObject selectButton;
+    public GameObject mainScreen;
+    public GameObject changeButton;
     public static PlayerSelector instance;
     public string playerPrefabName;
     public GameObject[] playerModel;
@@ -18,17 +22,35 @@ public class PlayerSelector : MonoBehaviour
 
     void Start()
     {
+        if (PlayerPrefs.HasKey("Gold"))
+        {
+            gold = PlayerPrefs.GetInt("Gold");
+        }
+        else
+        {
+            gold = 0;
+        }
+
         if (PlayerPrefs.HasKey("SelectedCharacter"))
         {
             selectedCharacter = PlayerPrefs.GetInt("SelectedCharacter");
             playerPrefabName = playerModel[selectedCharacter].GetComponent<PlayerModelName>().playerName;
             nextButton.SetActive(false);
             backwardButton.SetActive(false);
+            selectButton.SetActive(false);
+            mainScreen.SetActive(true);
+            changeButton.SetActive(true);
+
         }
         else
         {
+            selectedCharacter = 0;
+            playerPrefabName = playerModel[selectedCharacter].GetComponent<PlayerModelName>().playerName;
             nextButton.SetActive(true);
             backwardButton.SetActive(true);
+            selectButton.SetActive(true);
+            mainScreen.SetActive(false);
+            changeButton.SetActive(false);
         }
 
         foreach(GameObject player in playerModel)
@@ -46,7 +68,7 @@ public class PlayerSelector : MonoBehaviour
         if (selectedCharacter == playerModel.Length)
             selectedCharacter = 0;
         playerModel[selectedCharacter].SetActive(true);
-        PlayerPrefs.SetInt("SelectedCharacter", selectedCharacter);
+       
         playerPrefabName = playerModel[selectedCharacter].GetComponent<PlayerModelName>().playerName;
 
     }
@@ -57,13 +79,34 @@ public class PlayerSelector : MonoBehaviour
         if (selectedCharacter == -1)
             selectedCharacter = playerModel.Length -1;
         playerModel[selectedCharacter].SetActive(true);
-        PlayerPrefs.SetInt("SelectedCharacter", selectedCharacter);
+        
         playerPrefabName = playerModel[selectedCharacter].GetComponent<PlayerModelName>().playerName;
 
     }
 
-    public void Delete()
+    public void SelectChar()
     {
-        PlayerPrefs.DeleteKey("SelectedCharacter");
+        PlayerPrefs.SetInt("SelectedCharacter", selectedCharacter);
+        nextButton.SetActive(false);
+        backwardButton.SetActive(false);
+        selectButton.SetActive(false);
+        mainScreen.SetActive(true);
+        changeButton.SetActive(true);
+    }
+
+    public void ChangeChar(int amount)
+    {
+        if(gold >= amount)
+        {
+            gold -= amount;
+            PlayerPrefs.SetInt("Gold", gold);
+            PlayerPrefs.DeleteKey("SelectedCharacter");
+            nextButton.SetActive(true);
+            backwardButton.SetActive(true);
+            selectButton.SetActive(true);
+            mainScreen.SetActive(false);
+            changeButton.SetActive(false);
+        }
+       
     }
 }
